requirements.txt
asgiref==3.5.2
astroid==2.11.6
attrs==22.1.0
autobahn==22.6.1
Automat==20.2.0
autopep8==1.6.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.0
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
dill==0.3.5.1
Django==3.2
djangorestframework==3.13.1
fcm-django==1.0.12
firebase==3.0.1
firebase-admin==5.2.0
flake8==4.0.1
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
msgpack==1.0.4
pep8==1.7.1
Pillow==9.2.0
platformdirs==2.5.2
proto-plus==1.20.6
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pylint==2.14.4
pyOpenSSL==22.0.0
pyotp==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.28.1
rsa==4.9
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
wrapt==1.14.1
zope.interface==5.4.0
